requirements.txt
Flask==1.1.2
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
click==7.1.2
itsdangerous==1.1.0
PyYAML==5.3
pyecharts==1.8.1
pyecharts-snapshot==0.2.0